Refactor: regular chat api
